requirements.txt
alabaster==0.7.10
argh==0.26.2
Babel==2.5.3
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
defusedxml==0.5.0
Django==2.0.5
django-braces==1.13.0
django-multiselectfield==0.1.8
django-nose==1.4.6
django-oauth-toolkit==1.1.2
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==3.2.3
dnspython==1.16.0
docutils==0.14
entrypoints==0.3
Fedal-Restful-API==3.0
flake8==3.7.9
forex-python==1.0.0
httpie==0.9.9
idna==2.6
imagesize==1.0.0
itypes==1.1.0
Jinja2==2.10
livereload==2.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.19
mysqlclient==1.4.6
nose==1.3.7
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pathtools==0.1.2
pep8==1.7.1
port-for==0.4
protobuf==3.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
Pyliner==1.0
pyparsing==2.4.6
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.8
